% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstracts.R
\name{get_pubmed_article_abstract}
\alias{get_pubmed_article_abstract}
\alias{get_pubmed_article_conclusions}
\alias{get_pubmed_article_results}
\title{Retrieve Abstracts, Conclusions, and Results from PubMed Articles}
\usage{
get_pubmed_article_abstract(pubmed_article_element)

get_pubmed_article_conclusions(pubmed_article_element)

get_pubmed_article_results(pubmed_article_element)
}
\arguments{
\item{pubmed_article_element}{An XML element representing a PubMed article.}
}
\value{
A character string containing the abstract, conclusion, or result, depending on the function.
}
\description{
These internal functions retrieve abstracts, conclusions, and results from PubMed articles.
They are used to process the XML elements returned by the PubMed API.
}
\keyword{internal}
